Rename FakeHttpResponse to HttpResponseFake
