+++ src/ForEvolve.XUnit/Http/HttpContextHelper.cs
@@ -30,7 +30,7 @@ namespace ForEvolve.XUnit.Http
             HeaderDictionary = new HeaderDictionary();
             HttpRequest = new HttpRequestFake(HttpContextMock.Object, HeaderDictionary);
             ResponseCookiesMock = new Mock<IResponseCookies>();
-            HttpResponse = new FakeHttpResponse(
+            HttpResponse = new HttpResponseFake(
                 HttpContextMock.Object, 
                 HeaderDictionary, 
                 ResponseCookiesMock.Object
@@ -53,42 +53,5 @@ namespace ForEvolve.XUnit.Http
         }
 
         public HttpContextHelperOptions Options { get; set; }
-    }
-
-
-    public class FakeHttpResponse : HttpResponse
-    {
-        public FakeHttpResponse(HttpContext httpContext, IHeaderDictionary headers, IResponseCookies cookies)
-        {
-            HttpContext = httpContext;
-            Headers = headers;
-            Cookies = cookies;
-        }
-
-        public override HttpContext HttpContext { get; }
-        public override IHeaderDictionary Headers { get; }
-
-        public override int StatusCode { get; set; }
-        public override Stream Body { get; set; }
-        public override long? ContentLength { get; set; }
-        public override string ContentType { get; set; }
-
-        public override IResponseCookies Cookies { get; }
-        public override bool HasStarted { get; }
-
-        public override void OnCompleted(Func<object, Task> callback, object state)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override void OnStarting(Func<object, Task> callback, object state)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override void Redirect(string location, bool permanent)
-        {
-            throw new NotImplementedException();
-        }
     }
 }
